update teams page- move to a subpage
