+++ deployment/frontend/src/components/team/TeamHeaderCard.tsx
@@ -1,13 +1,17 @@
 import React from 'react'
 import Image from 'next/image'
 import { ChevronLeftIcon } from "@heroicons/react/20/solid";
+import { Button } from "../_shared/Button";
 
 export default function TeamHeaderCard() {
   return (
-    <section id="team-header-card" className=" w-full flex flex-col md:flex-row max-w-9xl mx-auto  font-acumin gap-y-4">
+    <section id="team-header-card" className=" w-full flex flex-col md:flex-row max-w-9xl mx-auto  font-acumin gap-y-4 sm:mt-12">
       <div className='relative w-full sm:px-6 md:px-0 md:w-[40%] xxl:w-[540px] h-[296px]'>
-        <div className='bg-white px-4 py-4 absolute top-[76%] '>
-          <div className='text-center px-3 py-3 bg-wri-gold rounded-sm font-bold text-base flex justify-center items-center gap-x-1'><ChevronLeftIcon className='w-5 h-5' /> <span>See all items</span></div>
+        <div className="absolute top-[78%] flex h-[68px] w-56 items-center justify-center rounded-t-[3px] bg-white ">
+          <Button>
+            <ChevronLeftIcon className="mb-1 mr-1 h-6 w-6" />
+            <span>See all topics</span>
+          </Button>
         </div>
         <div className='w-full h-full border-b-2 border-b-wri-green  shadow flex justify-center items-center'>
           <div className=' relative  w-[393px] h-[128px] '>
